[libpas] Implement Previous-Tag-Exclusion https://bugs.webkit.org/show_bug.cgi?id=302808 rdar://152167632
Reviewed by NOBODY (OOPS!).

Whenever we allocate an object, we first check the previous tag --
either through reliable inference, or if that is not possible, via
an LDG instruction to load the mtag directly. Then, we exclude this
previous-tag from the mask of valid tags thereafter passed in to the
IRG instruction used to generate the new tag for the incoming
allocation.
I chose not to give it its own enablement bit (rather tying it to
adjacent-tag-exclusion) because there's no case where we'd want to
have one of these enabled but not the other.