[Filters] Enable CoreGraphics filters for SVGFilters https://bugs.webkit.org/show_bug.cgi?id=302822 rdar://165057553
Reviewed by Nikolas Zimmermann.

We loop through all the effects in the SVGFilter graph and check whether each of
them supports CoreGraphics filters or not. If one of them does not support
CoreGraphics filters, we fall back to software filters.

The problem is SourceAlpha has to be in the SVGFilter graph regardless whether it
is used or not. And since SourceAlpha does not support CoreGraphics filters, we
always fallback to software filters.

The fix should be to loop through the effects of the SVGFilter expression.

* LayoutTests/css3/filters/effect-grayscale-hw.html:
* LayoutTests/css3/filters/effect-hue-rotate-hw.html:
* LayoutTests/css3/filters/effect-saturate-hw.html:
* LayoutTests/css3/filters/effect-sepia-hw.html:
* LayoutTests/css3/filters/svg-blur-filter-clipped.html:
* LayoutTests/imported/blink/css3/blending/svg-isolation-add-filter-expected.html:
* LayoutTests/imported/mozilla/svg/filters/feGaussianBlur-alpha-01.svg:
* LayoutTests/svg/filters/svg-gaussianblur-edgeMode-duplicate.svg:
* Source/WebCore/platform/graphics/cg/GraphicsContextCG.cpp:
(WebCore::GraphicsContextCG::setCGGaussianBlur):
* Source/WebCore/platform/graphics/filters/FEDropShadow.cpp:
(WebCore::FEDropShadow::supportedFilterRenderingModes const):
* Source/WebCore/platform/graphics/filters/FEGaussianBlur.cpp:
(WebCore::FEGaussianBlur::supportedFilterRenderingModes const):
* Source/WebCore/svg/graphics/filters/SVGFilterRenderer.cpp:
(WebCore::SVGFilterRenderer::supportedFilterRenderingModes const):