Web Inspector: avoid keeping `this` alive with DOM event listeners https://bugs.webkit.org/show_bug.cgi?id=302922
Reviewed by NOBODY (OOPS!).

In some cases, the GC might not be able to see that a DOM node, it's event listener, and the custom bound `this` object are an island that can be disposed.

As such, use the new `boundWeak` to wrap the callback given to all DOM event listeners so that `this` is held by a `WeakRef` instead.

This is overall very similar to <https://webkit.org/b/196956>.

Test: inspector/unit-tests/function-utilities.html