Teach Integer Range Optimization about unsigned comparisons https://bugs.webkit.org/show_bug.cgi?id=303205
Reviewed by NOBODY (OOPS!).

This is needed to make https://github.com/WebKit/WebKit/pull/54498
(Make for-of use single comparison) profitable.

In order to allow pruning to work, we introduce a new kind of relationship:

{ m_right } |- m_left > m_offset

This lets us conclude interesting things about m_left, while remembering
that this only holds as long as m_right is live.

This is important because if m_right is a shadow, then we must erase
all of its relationships when we execute an upsilon that writes to it.

We also therefore cannot use this relationship to prove any additional
relationships, since we have no way to hold a set of live nodes in a
relationship. That is fine though, since we basically only need this
to prove that an index is nonnegative anyway.