REGRESSION(303612@main): Links are not preserved as clickable annotations when exporting to PDF in Safari https://bugs.webkit.org/show_bug.cgi?id=303274 rdar://165575382
Reviewed by NOBODY (OOPS!).

The refactoring in 303612@main introduced a bug in outline painting of
inline content, resulting in PDF link annotations to be missing for <a>
elements.

Prior to the refactor, RenderInline::paintOutline() called hasOutline()
on RenderElement, rather than on RenderStyle. For links, the former
returns true when hasOutlineAnnotation() is true, even if there's no
`outline-style` set or there is a zero `outline-width`. This is not true
for the RenderStyle method, which prevented PDF annotations from being
added.

We fix by:
1. Using renderer.hasOutline() in the initial guard to handle links with
   outline annotations but no CSS outline style.
2. Adding hasThemedFocusRing check to match the focus ring logic in
   RenderInline::paintOutline() prior to 303612@main.
3. Restorng the early return by consulting RenderStyle::hasOutline()
   _after_ adding the annotation.

No new test, but TestWebKitAPI.PDFSnapshot.Links is now progressed.

* Source/WebCore/rendering/OutlinePainter.cpp:
(WebCore::OutlinePainter::paintOutline const):