[RenderStyleGen] Generate setHasExplicit* getter/setters https://bugs.webkit.org/show_bug.cgi?id=303276
Reviewed by NOBODY (OOPS!).

Extends generation of RenderStyle property getters and setters
to include the `setHasExplicit*` variants. This requires a few
new codegen options just like for the visited link variants:

    - render-style-has-explicitly-set-storage-container
    - render-style-has-explicitly-set-storage-name
    - render-style-has-explicitly-set-storage-path

To get all the `setHasExplicit*` variants, support for a few
missing properties, `color`, `direction`, and `writing-mode`
were added.

For `color`, the getter was marked as custom for now, as it
is explicitly exported, and the generator does not currently
support that. It's possible it would also be ok for it to be
made inline instead, but that can be determined in another
change.

For `direction`, and `writing-mode`, both the getters and
setters were marked as custom, and the getters were added
under the custom names `computedDirection` and `computedWritingMode`
to avoid conflicting with existing functions. Previously the
text of the getter was inlined in CSSProperties.json. The
setters needs to be custom due to each return a bool indicating
if the value changed, which is not something the generator can
do right now.

Additionally, all the `hasExplictlySet*` member names were
aligned and all the `hasExplictlySet*` setters were aligned
to take a bool argument.

* Source/WebCore/css/CSSProperties.json:
* Source/WebCore/css/scripts/process-css-properties.py:
* Source/WebCore/html/shadow/DateTimeFieldElement.cpp:
* Source/WebCore/rendering/style/RenderStyle.cpp:
* Source/WebCore/rendering/style/RenderStyle.h:
* Source/WebCore/rendering/style/RenderStyleInlines.h:
* Source/WebCore/rendering/style/RenderStyleSetters.h:
* Source/WebCore/rendering/style/StyleRareInheritedData.cpp:
* Source/WebCore/rendering/style/StyleRareInheritedData.h:
* Source/WebCore/style/StyleBuilderCustom.h: