[CSS Zoom] Remove unnecessary copy in conversionDataForStyle https://bugs.webkit.org/show_bug.cgi?id=303409 rdar://165711554
Reviewed by NOBODY (OOPS!).

The `conversionDataForStyle` method in RenderThemeCocoa.mm always makes a copy of
a new object just to change the value of two members. Since this object contains a
half dozen points, plus a few other members, this seems wasteful.

This patch simply sets the two members to the correct state when Evaluation Time
Zoom is enabled and avoids the copy.

This could be slightly improved by not setting the `m_zoom` member to 1.0f (since
that is the default constructed value in `CSSToLengthConversionData`, but I'm keeping
this assignment to keep the logic consistent with the old patch, and to avoid any issues
if the implementation of `CSSToLengthConversionData` every changes.

* Source/WebCore/rendering/cocoa/RenderThemeCocoa.mm:
(WebCore::conversionDataForStyle): Avoid copy.
* Source/WebCore/rendering/mac/RenderThemeMac.mm:
(WebCore::sizeForFont): Avoid extra copy.
(WebCore::sizeForSystemFont): Ditto.