[masonry] Fix sub grid tests row/grid-lanes-subgrid-001a and 002a https://bugs.webkit.org/show_bug.cgi?id=303500 rdar://problem/165791752
Reviewed by NOBODY (OOPS!).

grid-lanes-subgrid-001a.html
The 4c was placed below 4a instead of 4b, because the item-tolerance property defaults to 16px.
Changing to 0px will correctly set it below 4b.

grid-lanes-subgrid-002a.html
Heights were off because they were not being properly set. They are set by the grid-template-rows in the -ref case.
Needed to force item 4b to span across 2, because it is participating in the grid.

* LayoutTests/TestExpectations:
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-lanes/tentative/subgrid/row/grid-lanes-subgrid-001a.html:
* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/grid-lanes/tentative/subgrid/row/grid-lanes-subgrid-002a.html: