[JSC] Cache module evaluation errors to reject subsequent dynamic imports https://bugs.webkit.org/show_bug.cgi?id=303902
Reviewed by NOBODY (OOPS!).

According to the module evaluation spec[1], when a module is evaluated, a PromiseCapability representing
the evaluation result is set to module.[[TopLevelCapability]], and when the module is evaluated again, this
PromiseCapability is returned as-is.

In other words, module.[[TopLevelCapability]] acts as a cache. And according to the dynamic import spec[2],
dynamic import evaluates modules using the Evaluate Abstract Operation. So, in dynamic import, if a module
evaluation is rejected once, subsequent evaluations should also fail.

However, currently JSC resolves the second and subsequent evaluations even if the module was rejected once.

This patch fixes this by caching the evaluation error, similar to how link errors are cached, and re-throwing it
when the module is evaluated again.

[1]: https://tc39.es/ecma262/#sec-moduleevaluation
[2]: https://tc39.es/ecma262/#sec-ContinueDynamicImport

* JSTests/stress/re-execute-evaluation-error-module.js: Added.
(shouldBe):
(async let):
* JSTests/stress/resources/evaluation-error-module.js: Added.
* JSTests/test262/expectations.yaml:
* LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/choice-of-error-3-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/dynamic-import/dynamic-imports-script-error-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-error-1-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-error-2-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-error-3-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-error-4-expected.txt:
* Source/JavaScriptCore/builtins/ModuleLoader.js:
(linkTimeConstant.newRegistryEntry):
(visibility.PrivateRecursive.moduleEvaluation):
(visibility.PrivateRecursive.async asyncModuleEvaluation):