Web Inspector: Prevent automatic searches for potentially slow short queries https://bugs.webkit.org/show_bug.cgi?id=303939 rdar://166248017
Reviewed by NOBODY (OOPS!).

The previously used query is saved and automatically applied next time the Search tab is opened.
This was intentional, implemented in https://commits.webkit.org/184324@main

Web Inspector doesn't handle well short queries, like single characters, that return a lot of results.
The tool can slow down or even freeze. A user needs to close and reopen Web Inspector.

But the behaviour of automatically applying the last known query gets the user in the same state.
This patch prevents automatic searches for those potentially slow queries.
If the intent is real, pressing the Enter allows the user to proceed intentionally.

This is a stop-gap solution until adressing the underlying issue with slow queries.

* Source/WebInspectorUI/UserInterface/Views/SearchSidebarPanel.js:
(WI.SearchSidebarPanel.prototype.focusSearchField):