Trim whitespace from the end of keys for spell and grammar tests. https://bugs.webkit.org/show_bug.cgi?id=303966 rdar://166268991
Reviewed by NOBODY (OOPS!).

When writing tests for spelling and grammar checking, and there
is a newline as a part of the test, we can fail on iOS unless
the newline version is specified.
This is because of the differences between mac and iOS in how text
is iterated, and unless you happen to know this, or happen to look
at the right test to make it work, then you can spend a lot of
time trying to debug why the test isn't working.
If we just trim the whitespace off of these keys, then it's more
likely that someone in the future will not run into this issue.

* Tools/TestRunnerShared/cocoa/LayoutTestSpellChecker.mm:
(stringByTrimmingTrailingWhitespace):
(-[LayoutTestSpellChecker checkString:range:types:options:inSpellDocumentWithTag:orthography:wordCount:]):
(-[LayoutTestSpellChecker requestCheckingOfString:range:types:options:inSpellDocumentWithTag:completionHandler:]):
(-[LayoutTestSpellChecker requestGrammarCheckingOfString:range:language:options:completionHandler:]):
(-[LayoutTestSpellChecker requestProofreadingReviewOfString:range:language:options:completionHandler:]):
(swizzledGrammarDetailsForString):
(-[LayoutTestSpellChecker checkString:range:types:languages:options:]):