Add SPI to initialize _WKJSBuffer from NSString https://bugs.webkit.org/show_bug.cgi?id=303983 rdar://166287912
Reviewed by NOBODY (OOPS!).

Add SPI to initialize _WKJSBuffer from NSString. It's a frequent
use-case, since the buffer is used to transfer strings in some
APIs. It is a bit non-trivial to copy the string to the buffer and
handle the encodings correctly.

Similar to the NSData constructor, the SPI leaves abstract whether the
copy happens or whether the NSString is retained.

Test: Tools/TestWebKitAPI/Tests/WebKitCocoa/JSBuffer.mm
* Source/WebKit/UIProcess/API/Cocoa/_WKJSBuffer.h:
* Source/WebKit/UIProcess/API/Cocoa/_WKJSBuffer.mm:
(-[_WKJSBuffer initWithString:resultStringEncoding:]):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/JSBuffer.mm:
(TEST(JSBuffer, StringConstructor)):