Layout milestone notifications should be dispatched asynchronously https://bugs.webkit.org/show_bug.cgi?id=303958 rdar://153254633
Reviewed by NOBODY (OOPS!).

When a layout milestone notification fires, the layout may not be completed but client
callbacks to layout milestone notifications can trigger Javascript execution. This results
in javascript execution failing (and consequently crashing) to run during the
layout process because it is correctly not permitted to do so. We delay firing layout
milestone notifications by dispatching them asychronously in order to prevent this.

Test: Tools/TestWebKitAPI/Tests/WebKitCocoa/AutoLayoutIntegration.mm

* Source/WebCore/loader/FrameLoader.cpp:
(WebCore::FrameLoader::didReachLayoutMilestone):
* Source/WebCore/loader/LocalFrameLoaderClient.h:
* Source/WebKitLegacy/mac/WebCoreSupport/WebFrameLoaderClient.h:
* Source/WebKitLegacy/mac/WebCoreSupport/WebFrameLoaderClient.mm:
(WebFrameLoaderClient::willDispatchDidReachLayoutMilestone):
(WebFrameLoaderClient::dispatchDidReachLayoutMilestone):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/AutoLayoutIntegration.mm:
(TEST(WebKit, AutoLayoutBatchesUpdatesWhenInvalidatingIntrinsicContentSize)):