`JSON.parse` could allocate Array backing stores uninitialized https://bugs.webkit.org/show_bug.cgi?id=304264 rdar://164716376
Reviewed by NOBODY (OOPS!).

Instead of constructEmptyArray in LiteralParser::parseRecursively, this patch
uses JSArray::tryCreateUninitializedRestricted with a GCDeferralContext.
This avoids the overhead of zeroing out the array's backing store.

This is a performance change that's covered by existing tests,
such as JetStream's json-parse-inspector benchmark.

* Source/JavaScriptCore/runtime/LiteralParser.cpp:
(JSC::LiteralParser::parseRecursively): replaces usage of constructEmptyArray with JSArray::tryCreateUninitializedRestricted