Make height-dependent aspect-ratio set width on tables with specified height https://bugs.webkit.org/show_bug.cgi?id=304709 rdar://167214714
Reviewed by NOBODY (OOPS!).

In RenderTable::layout() we weren't setting the computed height onto
m_frameRect before calling updateLogicalWidth(), but updateLogicalWidth()
passes the m_frameRect's logicalHeight() as the definite height for
calculating a height-dependent width via the aspect-ratio in
RenderBox::computeLogicalWidthFromAspectRation(). This resulted in us
always calculating an aspect-ratio dependent width of zero.

With this change, if there's an aspect-ratio, we compute the height and
store it onto m_frameRect before calling updateLogicalWidth() so that
an aspect-ratio-dependent width computes correctly.

* LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-item-aspect-ratio-justify-self-002-expected.txt:
* Source/WebCore/rendering/RenderTable.cpp:
(WebCore::RenderTable::layout):