Preserve the exact color values defined by keyframes when animating colors https://bugs.webkit.org/show_bug.cgi?id=304764
Reviewed by NOBODY (OOPS!).

When animating colors, we now special case progress=0% and progress=100%
to return the unmodified `from` and `to` colors respectively, rather than
the old behavior, where the `from` and `to` colors would be returned as
values converted to the interpolation color space (Oklab).

See https://github.com/w3c/csswg-drafts/issues/13269 for more.

We could consider implementing this in `Style::Interpolation` and applying
it for all properties, but so far, only colors need to be special cased,
as other properties don't have a similar "conversion to interpolation color
space" concept and thus the extra branches would be wasted.

Test: imported/w3c/web-platform-tests/css/css-animations/animate-color-preserve-exact-serialization-at-keyframes.html
* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/animate-color-preserve-exact-serialization-at-keyframes-expected.txt: Added.
* LayoutTests/imported/w3c/web-platform-tests/css/css-animations/animate-color-preserve-exact-serialization-at-keyframes.html: Added.
* Source/WebCore/platform/graphics/ColorBlending.cpp: